+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - metrics.clear() disables default metrics
 
+### Added
+- Luajit platform metrics
+
 ## [0.5.0] - 2020-09-18
 ### Added
 - Summary collector