+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - metrics.clear() disables default metrics
+- cartridge role is permanent
 
 ## [0.5.0] - 2020-09-18
 ### Added