+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Luajit platform metrics
 
+### Changed
+- CI on Github Actions
+
 ## [0.5.0] - 2020-09-18
 ### Added
 - Summary collector