+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2020-12-01
 ### Fixed
 - metrics.clear() disables default metrics
 - cartridge role is permanent
@@ -12,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - graphite plugin kills previous workers on init
 - graphite plugin format numbers without ULL/LL-suffixes
 - graphite plugin time in seconds
+- graphite plugin allows empty prefix
 
 ### Added
 - Luajit platform metrics